+++ src/core/encoder/encoders/avi_encoder.ml
@@ -116,7 +116,7 @@ let encoder ~pos:_ avi =
     Encoder.insert_metadata = (fun _ -> ());
     hls = Encoder.dummy_hls encode;
     encode;
-    header = Strings.of_string header;
+    header = (fun () -> Strings.of_string header);
     stop = (fun () -> Strings.empty);
   }