+++ CHANGES.md
@@ -5,6 +5,8 @@ Fixed:
 - Fixed HTTP response status in `output.harbor` (#3255)
 - Make sure main HLS playlist is regenerated after being
   unlinked (#3275)
+- Fix `output.harbor` encoder header when encoding with
+  `%ogg`, `%vorbis` and etc. (#3276)
 
 ---